clean up + renamed duplicated file names to avoid confusion
