+++ src/languageservice/services/yamlHover.ts
@@ -13,7 +13,7 @@ import { isAlias, isMap, isSeq, Node, stringify as stringifyYAML } from 'yaml';
 import { ASTNode, ObjectASTNode, PropertyASTNode } from '../jsonASTTypes';
 import { JSONSchema } from '../jsonSchema';
 import { setKubernetesParserOption } from '../parser/isKubernetes';
-import { getNodeValue } from '../parser/astUtils';
+import { getNodeValue } from '../parser/astNodeUtils';
 import { IApplicableSchema } from '../parser/schemaValidation/baseValidator';
 import { yamlDocumentsCache } from '../parser/yaml-documents';
 import { SingleYAMLDocument } from '../parser/yamlParser07';
