+++ src/languageservice/utils/yamlAstUtils.ts
