+++ test/astUtils.test.ts
@@ -6,7 +6,7 @@
 import * as chai from 'chai';
 import { isPair, isSeq, Pair, YAMLSeq } from 'yaml';
 import { YamlDocuments } from '../src/languageservice/parser/yaml-documents';
-import { getParent, isInComment } from '../src/languageservice/utils/astUtils';
+import { getParent, isInComment } from '../src/languageservice/utils/yamlAstUtils';
 import { TextBuffer } from '../src/languageservice/utils/textBuffer';
 import { setupTextDocument } from './utils/testHelper';
 const expect = chai.expect;
