+++ test/jsonParser.test.ts
@@ -6,7 +6,7 @@
 
 import * as assert from 'assert';
 import { JSONDocument } from '../src/languageservice/parser/jsonDocument';
-import { getNodeValue } from '../src/languageservice/parser/astUtils';
+import { getNodeValue } from '../src/languageservice/parser/astNodeUtils';
 import * as JsonSchema from './../src/languageservice/jsonSchema';
 import { ASTNode, ObjectASTNode } from './../src/languageservice/jsonASTTypes';
 import { ErrorCode, getLanguageService } from 'vscode-json-languageservice';
